service-authentication/requirements.txt
alembic==1.14.0
arrow==1.3.0
blinker==1.9.0
cleo==2.1.0
crashtest==0.4.1
dsnparse==0.2.1
greenlet==3.1.1
grpcio==1.69.0
grpcio-tools==1.69.0
inflection==0.5.1
lazy-object-proxy==1.10.0
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.25.1
orator==0.6.4
packaging==24.2
protobuf==5.29.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
RapidFuzz==3.11.0
setuptools==75.8.0
simplejson==3.19.3
six==1.17.0
SQLAlchemy==2.0.37
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
pyotp
PyJWT
python-dotenv